refactor(delivery): delete retired fanout runtime
