+++ internal/access/node/delivery_rpc.go
@@ -13,9 +13,6 @@ import (
 // DeliveryPushRPCServiceID is the cluster RPC service for owner-node delivery batches.
 const DeliveryPushRPCServiceID uint8 = clusternet.RPCDeliveryPush
 
-// DeliveryFanoutRPCServiceID is the cluster RPC service for authority-node fanout tasks.
-const DeliveryFanoutRPCServiceID uint8 = clusternet.RPCDeliveryFanout
-
 // HandleDeliveryPushRPC handles one encoded delivery push RPC payload.
 func (a *Adapter) HandleDeliveryPushRPC(ctx context.Context, payload []byte) ([]byte, error) {
 	req, err := decodeDeliveryPushRequest(payload)
@@ -45,37 +42,6 @@ func (a *Adapter) HandleDeliveryPushRPC(ctx context.Context, payload []byte) ([]
 		return encodeDeliveryPushResponse(deliveryPushResponse{Status: rpcStatusRejected})
 	}
 	return encodeDeliveryPushResponse(deliveryPushResponse{Status: rpcStatusOK, Result: result})
-}
-
-// HandleDeliveryFanoutRPC handles one encoded delivery fanout task RPC payload.
-func (a *Adapter) HandleDeliveryFanoutRPC(ctx context.Context, payload []byte) ([]byte, error) {
-	req, err := decodeDeliveryFanoutRequest(payload)
-	if err != nil {
-		a.rpcLogger().Warn("delivery fanout rpc decode failed",
-			wklog.Event("internal.access.node.delivery_fanout_decode_failed"),
-			wklog.Int("payloadBytes", len(payload)),
-			wklog.Error(err),
-		)
-		return nil, err
-	}
-	if a == nil || a.deliveryFanout == nil {
-		return encodeDeliveryFanoutResponse(deliveryFanoutResponse{Status: rpcStatusRejected})
-	}
-	if err := a.deliveryFanout.RunTask(ctx, req.Task); err != nil {
-		a.rpcLogger().Warn("delivery fanout rpc rejected",
-			wklog.Event("internal.access.node.delivery_fanout_rejected"),
-			wklog.Uint64("targetNodeID", req.Task.Partition.LeaderNodeID),
-			wklog.Int("partitionID", int(req.Task.Partition.ID)),
-			wklog.ChannelID(req.Task.Envelope.ChannelID),
-			wklog.ChannelType(int64(req.Task.Envelope.ChannelType)),
-			wklog.Uint64("messageID", req.Task.Envelope.MessageID),
-			wklog.MessageSeq(req.Task.Envelope.MessageSeq),
-			wklog.Attempt(req.Task.Attempt),
-			wklog.Error(err),
-		)
-		return encodeDeliveryFanoutResponse(deliveryFanoutResponse{Status: rpcStatusRejected})
-	}
-	return encodeDeliveryFanoutResponse(deliveryFanoutResponse{Status: rpcStatusOK})
 }
 
 // PushBatch forwards one owner-node delivery batch to nodeID.
@@ -113,31 +79,4 @@ func (c *Client) PushOwner(ctx context.Context, cmd onlinedelivery.OwnerPush) (o
 		return onlinedelivery.OwnerPushResult{}, err
 	}
 	return onlineDeliveryResultFromLegacy(result), nil
-}
-
-// ForwardFanoutTask forwards one partition-scoped fanout task to nodeID.
-func (c *Client) ForwardFanoutTask(ctx context.Context, nodeID uint64, task runtimedelivery.FanoutTask) error {
-	if c == nil || c.node == nil {
-		return fmt.Errorf("internal/access/node: delivery fanout rpc client not configured")
-	}
-	body, err := encodeDeliveryFanoutRequest(deliveryFanoutRequest{Task: task})
-	if err != nil {
-		return err
-	}
-	respBody, err := c.node.CallRPC(ctx, nodeID, DeliveryFanoutRPCServiceID, body)
-	if err != nil {
-		return err
-	}
-	resp, err := decodeDeliveryFanoutResponse(respBody)
-	if err != nil {
-		return err
-	}
-	switch resp.Status {
-	case rpcStatusOK:
-		return nil
-	case rpcStatusRejected:
-		return fmt.Errorf("internal/access/node: delivery fanout rpc rejected")
-	default:
-		return fmt.Errorf("internal/access/node: unknown delivery fanout rpc status %q", resp.Status)
-	}
 }
